requirements.txt
pytest~=9.0.2
allure-python-commons~=2.15.3
openai~=2.14.0
requests~=2.32.5
pandas~=2.3.3
openpyxl~=3.1.5
PyYAML~=6.0.3
pytest-xdist==3.3.1
pytest-rerunfailures==12.0
allure-pytest==2.13.2
Jinja2~=3.1.6
langfuse~=3.12.0